requirements.txt
aiogram>=3.4.0
aiosqlite>=0.19.0
SQLAlchemy[asyncio]>=2.0.25
alembic>=1.13.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
Pillow>=10.0.0
